openconnect: fix do_compile failure
Issue: LIN9-538

This patch solves the following error which appears now and then because of
parallel building.

  ../git/main.c:78:21: fatal error: version.c: No such file or directory

The problem is that version.c needs to be there when building some objects
which needs main.c. But version.c depends on main.c and is generated at
build time.

This patch fix the problem by generating version.c first before building others.

(LOCAL REV: NOT UPSTREAM -- sent to oe on 20160512)

Signed-off-by: Chen Qi <Qi.Chen@windriver.com>
Signed-off-by: Robert Yang <liezhi.yang@windriver.com>