lmsensors: read cpu information fail on ppc
Issue: LIN9-1576

The program sensors-detect in lmsensors can't read the cpu
information on ppc and arm arch, such as: fsl-t4xxx based on ppc.

This is because the program sensors-detect design for the x86 and
x86-64, but weak on ppc and arm.

So add the support for the ppc and arm, just add the correspond
field to read in sensors-detect.

(LOCAL REV: NOT UPSTREAM) -- Send to meta-openembedded 20160922

Signed-off-by: Dengke Du <dengke.du@windriver.com>