apache2: include .load files in modules.d
Issue: LIN9-1923

According to other linux distributes like ubuntu, the modules
are usually included by 'LoadModule' command in *.load files
in mods-enable directory, and *.conf files in this directory
are for special configurations for each module.

The stucture of the files tree is not the same in yocto, so this
may be not the rule that we must follow. But still add *.load to
be compatible with customer's normal usage habits.

(LOCAL REV: NOT UPSTREAM) -- sent to oe-devel on 20161031

Signed-off-by: Junxian.Xiao <Junxian.Xiao@windriver.com>
Signed-off-by: Kai Kang <kai.kang@windriver.com>