curlpp: fix qa issue of do_configure
Issue: LIN9-2204

curlpp is configured with boost by default. And it searches boost
directory on build machine. If boost is installed on build machine, then
causes qa issue:

| ERROR: curlpp-0.7.3-r0 do_configure: This autoconf log indicates
| errors, it looked at host include and/or library paths while
| determining system capabilities.

Pass the boost directory in ${STAGING_DIR_HOST} to configure script to
disable search on build machine to fix the issue.

(LOCAL REV: NOT UPSTREAM) -- Send to oe-devel on 20161116

Signed-off-by: Kai Kang <kai.kang@windriver.com>