mozjs: exclude mips64 from COMPATIBLE_HOST
Issue: LIN9-2369

mozjs 17 doesn't support mips64. mips64 support has been added on Nov 7,
2015 by

https://github.com/mozilla/gecko-dev/commit/8b4f672

Set COMPATIBLE_HOST for mozjs and packages which depend on it directly
or indirectly.

(LOCAL REV: NOT UPSTREAM) -- Won't send to oe-devel that mozjs should be
upgraded to latest version 45 which supports mips64. This commit should
be reverted for WRL10.

Signed-off-by: Kai Kang <kai.kang@windriver.com>