rsyslog: use atomic builtins to avoid race issue
Issue: LIN9-1953

rsyslog uses a small program to check if compiler has atomic builtins
or not. This does not work when cross compiling, resulting in potential
race issue. In fact, when running ptest on intel-x86-64 board, we have
met the race problem.

As our compiler gcc, has atomic builtins, we force ap_cv_atomic_builtins
to be 'yes' so that rsyslog could make use of the atomic builtins and avoid
the race issue.

(LOCAL REV: NOT UPSTREAM) -- sent to oe on 20161228

Signed-off-by: Chen Qi <Qi.Chen@windriver.com>