makedumpfile: update the latest kernel version
Issue: CGP9-95
Issue: TA171134

It defines the latest kernel version as 4.8.1, but the wrlinux9 kernel
version is 4.8.17 which cause the warning:
The kernel version is not supported.
The makedumpfile operation may be incomplete.

Actually it works well on this kernel, so update the latest kernel
version to 4.8.17.

(LOCAL REV: NOT UPSTREAM) -- will send to oe-devel on 20170210

Signed-off-by: Yi Zhao <yi.zhao@windriver.com>