squid: fix do_compile with gcc7
Issue: LIN10-695

After gcc upgrade to 7.1, gcc7 contains a number of enhancements that
help detect buffer overflow and other forms of invalid memory accesses.

    https://gcc.gnu.org/gcc-7/changes.html

When the squid compiled with gcc7, it outputs two errors:

    Werror=format-truncation=
    Werror=int-in-bool-context=

This is because the squid's Makefile set the -Werror, so the gcc treat
the warnings as errors, in this way, we should remove it.

(LOCAL REV: NOT UPSTREAM) -- Send to meta-networking on 20170626

Signed-off-by: Dengke Du <dengke.du@windriver.com>