python-twisted: restore RDEPENDS on -news for py2
Although this plugin has not yet been ported to py3 it is available in
py2. The previous commit isn't clear about what is being addressed but
I assume it is:

nothing provides python3-twisted-news needed by python3-twisted-17.9.0-r0.corei7_64

For py2 this package is valid and so we should maintain the RDEPENDS,
restore this RDEPENDS for py2.

(LOCAL REV: NOT UPSTREAM) Sent upstream 15 Dec 2017

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>