gstreamer: Remove machine specific append
If you try to build a system with multiple BSPs, one of which is qemux86
or qemux86-64, the gstreamer package will change.  This will trigger
anything using gstream to also be rebuilt.

For a package based system, the PR values will also be incremented each
time.  The end result will be an ever growing set of PR values as well as
being unable to tell which configured version of the multimedia components
are really being deployed.

The solution here was to remove the rrecommend for consistency.

(LOCAL REV: NOT UPSTREAM) -- Sent to openembedded-devel on 20181017

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>