pmdk: force libraries to be installed to lib64
Issue: LIN1018-3419

pmdk libraries can be installed to lib or lib64, depending on
host building system. If host has /usr/local/lib64, libraries will be
installed in lib64, otherwise they will be installed in lib.
pmdk is only for 64-bit system, and to make the install consistent,
force the install path to be lib64.

(LOCAL REV: NOT UPSTREAM) -- Not Applicable

Signed-off-by: Yongxin Liu <yongxin.liu@windriver.com>