crda: wireless-regdb -> wireless-regdb-static
The wireless-regdb has been moved to oe-core, and according the commit message:

wireless-regdb-static should be used with kernel >= 4.15.
wireless-regdb can be used with older kernels and is mostly
irrelevant here, but keeping it in meta-networking would
create needless recipe duplication.

So use wireless-regdb-static to replace of wireless-regdb

(LOCAL REV: NOT UPSTREAM) -- Sent to meta-oe on 20190625

Signed-off-by: Robert Yang <liezhi.yang@windriver.com>