strongswan: fix CVE-2021-41990 and CVE-2021-41991
Issue: LIN1021-1737
Issue: LIN1021-1738

CVE-2021-41990:
The gmp plugin in strongSwan before 5.9.4 has a remote integer overflow
via a crafted certificate with an RSASSA-PSS signature. For example,
this can be triggered by an unrelated self-signed CA certificate sent by
an initiator. Remote code execution cannot occur.

CVE-2021-41991:
The in-memory certificate cache in strongSwan before 5.9.4 has a remote
integer overflow upon receiving many requests with different
certificates to fill the cache and later trigger the replacement of
cache entries. The code attempts to select a less-often-used cache entry
by means of a random number generator, but this is not done correctly.
Remote code execution might be a slight possibility.

References:
https://nvd.nist.gov/vuln/detail/CVE-2021-41990
https://nvd.nist.gov/vuln/detail/CVE-2021-41991

Patches from:
https://github.com/strongswan/strongswan/commit/234302a108b490635fb961ebd1404340c46cefc3
https://github.com/strongswan/strongswan/commit/dceed8e099c8f9d5606fa7e5a0742fb5e023103b

(LOCAL REV: NOT UPSTREAM) -- The upstream version 5.9.4 is not affected

Signed-off-by: Yi Zhao <yi.zhao@windriver.com>