plymouth: switch to KillMode=mixed
Issue: LIN1021-2478

KillMode=none is deprecated, so we need to stop using it [1].

For now, use `KillMode=mixed` and `IgnoreOnIsolate=true` instead.

In the future, we should change plymouth to be able to exit and
start again without restarting the active animation, but that's
going to require some effort.

[1] https://gitlab.freedesktop.org/plymouth/plymouth/-/issues/123

(LOCAL REV: NOT UPSTREAM) -- Sent to oe-devel on 20220124

Signed-off-by: Mingli Yu <mingli.yu@windriver.com>