lxdm: remove conflicts setting
Issue: LIN1021-2592 LINUXEXEC-18176

When use plymouth[1] with lxdm, the plymouth boot splash screen
will hang there as lxdm.service conflicts with plymouth-quit.service
and plymouth-quit will fail to start if plymouth-quit.service and
lxdm.service start at the same time and it will result in the boot
screen hang forever.

Consider there is also After setting for these two services, so
remove the conflicts setting to make the service start as expected.

[1] https://gitlab.freedesktop.org/plymouth/plymouth/-/tree/main

(LOCAL REV: NOT UPSTREAM) -- Sent to oe-devel on 20220125

Signed-off-by: Mingli Yu <mingli.yu@windriver.com>